cli/src/dev: move phpmyadmin under 'dev' group
